day-13-py/pyproject.toml
[tool.poetry]
name = "day13py"
version = "0.0.1"
description = ""
authors = ["Denis Viviès <dvivies@bmat.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.1"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"
httpx = "^0.23.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
mypy = "^0.950"

[tool.pytest.ini_options]
addopts = "-ra --cov=day13py"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
